+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 ## Changelog
 
+### 1.1.7
+
+*2018-03-31*
+
+- Новая стилизация simplified компонента Dialog
+- Добавлены новые иконки:
+    - transfer
+    - umbrella (старый заменен на umbrella-angle)
+    - aeroexpress
+
 ### 1.1.6
 
 *2018-03-30*